requirements.txt
Flask==0.12.2
matplotlib==2.0.0
numpy==1.12.0
Pillow==4.2.1
python-dateutil==2.6.0
six==1.10.0
waitress==1.0.2
Werkzeug==0.12.2
tensorflow==1.2.1